Barely working version
This "works", but it does so by popping up the mobile version of Remember the
Milk web site and using that to send the task to Remember the Milk.

This should use the Remember the Milk API, but I need something that works now
even if it's mediocre. We can switch to the API in a future version.